breakup Tmux notifier into smaller files
